make stage block a one of block
