+++ api/src/documents/pages/blocks/basic-stage.block.ts
@@ -1,0 +1,48 @@
+import {
+    BlockData,
+    BlockDataInterface,
+    BlockInput,
+    ChildBlock,
+    ChildBlockInput,
+    createBlock,
+    ExtractBlockInput,
+    inputToData,
+} from "@comet/blocks-api";
+import { CallToActionListBlock } from "@src/common/blocks/call-to-action-list.block";
+import { HeadingBlock } from "@src/common/blocks/heading.block";
+import { MediaBlock } from "@src/common/blocks/media.block";
+import { RichTextBlock } from "@src/common/blocks/rich-text.block";
+
+class BasicStageBlockData extends BlockData {
+    @ChildBlock(MediaBlock)
+    media: BlockDataInterface;
+
+    @ChildBlock(HeadingBlock)
+    heading: BlockDataInterface;
+
+    @ChildBlock(RichTextBlock)
+    text: BlockDataInterface;
+
+    @ChildBlock(CallToActionListBlock)
+    callToActionList: BlockDataInterface;
+}
+
+class BasicStageBlockInput extends BlockInput {
+    @ChildBlockInput(MediaBlock)
+    media: ExtractBlockInput<typeof MediaBlock>;
+
+    @ChildBlockInput(HeadingBlock)
+    heading: ExtractBlockInput<typeof HeadingBlock>;
+
+    @ChildBlockInput(RichTextBlock)
+    text: ExtractBlockInput<typeof RichTextBlock>;
+
+    @ChildBlockInput(CallToActionListBlock)
+    callToActionList: ExtractBlockInput<typeof CallToActionListBlock>;
+
+    transformToBlockData(): BasicStageBlockData {
+        return inputToData(BasicStageBlockData, this);
+    }
+}
+
+export const BasicStageBlock = createBlock(BasicStageBlockData, BasicStageBlockInput, "BasicStage");
